requirements.txt
absl-py==2.1.0
accelerate==0.34.2
adapter==0.1
adapter-transformers==4.0.0
adapters==1.0.0
alembic==1.14.0
astunparse==1.6.3
brisque==0.0.16
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
easyocr==1.7.2
editdistance==0.8.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h5py==3.12.1
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
Levenshtein==0.26.0
libclang==18.1.1
libsvm-official==3.35.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
optuna==4.1.0
packaging==24.1
peft==0.13.0
pillow==10.4.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
Pygments==2.18.0
pyparsing==3.1.4
python-bidi==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tesserocr==2.7.1
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.43.4
triton==3.0.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
